+++ geonamescache/types.py
@@ -9,6 +9,8 @@ else:
 GeoNameIdStr = str
 ISOStr = str
 Admin1CodeStr = str
+Admin2CodeStr = str
+TimeZoneIdStr = str
 ContinentCode = Literal["AF", "AN", "AS", "EU", "NA", "OC", "SA"]
 USStateCode = Literal[
     "AK",
@@ -116,13 +118,21 @@ USStateName = Literal[
     "West Virginia",
     "Wyoming",
 ]
-CitySearchAttribute = Literal["alternatenames", "admin1code", "countrycode", "name", "timezone"]
+CitySearchAttribute = Literal["alternatenames", "admin1code", "admin2code", "countrycode", "name", "timezone"]
 
 
 class TimeZone(TypedDict):
     dstOffset: int
     gmtOffset: int
     timeZoneId: str
+
+
+class TimeZoneInfo(TypedDict):
+    countrycode: str
+    timezoneid: str
+    gmtoffset: float
+    dstoffset: float
+    rawoffset: float
 
 
 class BBox(TypedDict):
@@ -174,9 +184,16 @@ class Admin1(TypedDict):
     name: str
 
 
+class Admin2(TypedDict):
+    asciiname: str
+    geonameid: int
+    name: str
+
+
 class City(TypedDict):
     alternatenames: list[str]
     admin1code: str
+    admin2code: str
     countrycode: str
     geonameid: int
     latitude: float